requirements.txt
Django=1.6
django-appconf=0.6
django-statici18n=1.1
djangorestframework=2.3
django-filter=0.7
django-angular=0.7